Added README
This initial commit includes the README file that provides an overview of the repository's purpose, its contents, and usage guidelines. The repository is set up to showcase coding assignments and projects completed during the COMP361 Data Analytics with Python course. The README serves as a helpful introduction for visitors and users of the repository.